+++ src/ainter/models/nagel_schreckenberg/road.py
@@ -38,7 +38,7 @@ class Road:
                                                                      [end_node_info['x'], end_node_info['y']]]))
 
         return cls(osm_id=osm_id,
-                   grid=np.zeros(shape=(cells_num, lanes), dtype=np.uint8),
+                   grid=np.zeros(shape=(cells_num, lanes), dtype=np.uint16),
                    lanes=lanes,
                    max_speed=max_speed,
                    name=name,
